pyproject.toml
[tool.poetry]
name = "chatstory"
version = "0.1.0"
description = "Project with FastAPI"
authors = ["junyong"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
pydantic = "<2.0"
pymongo = "^4.11.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
rq = "^2.3.2"
python-jose = "^3.4.0"
cryptography = "^44.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
